--- instrumentation/runtime/crash/src/main/kotlin/com/splunk/rum/instrumentation/crash/CrashDetails.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright 2026 Splunk Inc.
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.splunk.rum.instrumentation.crash
-
-/**
- * Holds the details of an application crash.
- *
- * @property thread The thread that crashed.
- * @property cause The throwable that caused the crash.
- */
-data class CrashDetails(val thread: Thread, val cause: Throwable)
