+++ integration/slowrendering/src/main/kotlin/com/splunk/rum/integration/slowrendering/SlowRenderingModuleIntegration.kt
@@ -18,7 +18,7 @@ package com.splunk.rum.integration.slowrendering
 
 import android.content.Context
 import com.splunk.android.common.logger.Logger
-import com.splunk.rum.instrumentation.slowrendering.SlowRenderingInstrumentation
+import com.splunk.rum.instrumentation.slowrendering.internal.SlowRenderingInstrumentation
 import com.splunk.rum.integration.agent.common.module.ModuleConfiguration
 import com.splunk.rum.integration.agent.common.module.find
 import com.splunk.rum.integration.agent.internal.legacy.LegacySlowRenderingModuleConfiguration
